requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tensorboard>=2.10.0
wandb>=0.13.0
pyyaml>=6.0
tqdm>=4.64.0
pandas>=1.3.0
disentanglement-lib>=1.0.0
pytorch-lightning>=1.9.0
omegaconf>=2.3.0
hydra-core>=1.3.0
